ml_backend/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
python-dotenv
requests
google-genai
